requirements.txt
requests
pandas
google-cloud-storage
lxml
mysql-connector-python
datetime

